JS To-Do List
This is a simple to-do list that I created using JavaScript. The styling of the to-do list was completed using Tailwind CSS.

This project helped me realize how small my knowledge in JS really was. However, I used my resources and was able to learn a lot during this process. Hopefully this list helps!